+++ src/enums.ts
@@ -56,6 +56,7 @@ export enum ScoreName {
     FAVOURITED_TAGS = 'FavouritedTags',
     FOLLOWED_ACCOUNTS = 'FollowedAccounts',
     FOLLOWED_TAGS = 'FollowedTags',
+    // FOLLOWERS = 'Followers',
     IMAGE_ATTACHMENTS = 'ImageAttachments',
     INTERACTIONS = 'Interactions',
     MENTIONS_FOLLOWED = 'MentionsFollowed',