+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
-* Add getFollowers() to MastoApi; add configuration variable config.api.pullFollowers to turn on/off follower retrieval
+* Add `getFollowers()` to `MastoApi`; add configuration variable `config.api.pullFollowers` to turn on/off follower retrieval
+* Move `TheAlgorithm.filterOptionDataSources()` to `BooleanFilter` static method
 
 ## v0.67.0
 * Add switches in the hashtag filter section to enable / disable each kind of hashtag colored highlighting (participated/favourited/trending)